backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
dropbox==11.36.2
idna==3.4
inflection==0.5.1
markdown2==2.4.10
packaging==23.2
Pillow==10.1.0
ply==3.11
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
dropbox==11.36.2
idna==3.4
inflection==0.5.1
markdown2==2.4.10
packaging==23.2
Pillow==10.1.0
ply==3.11
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
